refactor(cname-prefix): the hash back in its own file
`sha256EncodeSync.ts` held the compression function, the round constants and
the bigint conversion, which is two jobs in one file. `sha256.ts` now owns the
algorithm and returns the eight digest words; `sha256EncodeSync` only folds
them into the bigint the prefix is derived from, the same split the file had
when the implementation was vendored.

Also drops the AI tells from the README's Usage section: mechanical boldface
on four consecutive paragraphs, "the honest summary is", two manufactured
closers and every em dash.
